backend/requirements.txt
django==4.2
djangorestframework==3.16
python-decouple==3.8
pillow==11.1.0
hashids==1.3.1
drf-extra-fields==3.7.0
django-filter==25.1
python-docx==1.1.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-cleanup==9.0.0

